CmsAdmin/requirements.txt
aniso8601==7.0.0
appdirs==1.4.4
asgiref==3.2.10
black==20.8b1
certifi==2020.6.20
click==7.1.2
Django==2.2.16
django-cors-headers==3.5.0
django-jet==1.0.8
ecdsa==0.14.1
feedparser==6.0.1
graphene==2.1.8
graphene-django==2.13.0
graphql-core==2.3.2
graphql-relay==2.0.1
mypy-extensions==0.4.3
pathspec==0.8.0
promise==2.3
pyasn1==0.4.8
python-jose==3.2.0
pytz==2020.1
regex==2020.7.14
rsa==4.6
Rx==1.6.1
sgmllib3k==1.0.0
singledispatch==3.4.0.3
six==1.15.0
sqlparse==0.3.1
toml==0.10.1
typed-ast==1.4.1
typing-extensions==3.7.4.3
Unidecode==1.1.1
gunicorn==20.0.4
grpcio==1.34.0
grpcio-tools==1.34.0
djangorestframework==3.8.0
mysqlclient==1.4.0
django-filter==2.0.0
maya==0.6.1